Basic function arity support
Function signature resolution is pretty easy to do when using argument count, because that value is always known at compile time. But in order to do this during runtime based on types (unless those types are also known at compile time), we would need to generate wrappers for each function in the jsCompiler. These wrappers would check types and determine the function signature based on those types. This would necessarily mean combining functions into one unit in the JS Compiler, which means enforcing some sort of rules about scope. A sane rule might be that function definitions in the same module must happen consecutively and in the same scope.